refactor: extract components from index.tsx into named files
Move component logic out of index.tsx files into properly named files
(e.g., Sidebar/index.tsx → Sidebar/Sidebar.tsx) with index.tsx becoming
a pure barrel re-export. This enforces the convention that index files
are always barrels, never component definitions.

Extracted:
- Sidebar (532 lines)
- Grid (476 lines)
- BinListModal (455 lines)
- LayoutManagerModal (308 lines)
- MobileLayersPanel (38 lines)

Also split PrintListComponents.test.tsx (loose multi-component test)
into colocated PrintListEmpty.test.tsx — the other tests already
existed as colocated files.
